+++ DICs_API/DICs_API/Models/Status.cs
@@ -11,6 +11,6 @@ namespace DICs_API.Models
         public int Id { get; set; }
         [Required]
         public string Name { get; set; }
-        public int Removed { get; set; }
+        public byte Removed { get; set; }
     }
 }
